fix: adjusting build issues
